wip: seperated weight systems from the main codebase
